Added Script for WIN NT Clients.
Same functionality as the batch version - will check if venv is initialized and if not will do so, then start the Flask server. Currently not using a WSGI server. Not super neccesary but might as well at some point?

+++ lunar.bat
@@ -1,6 +1,14 @@
-python3 -m venv .lenv
-.lenv\Scripts\activate.bat
-pip install -r requirements.txt
+%~dp0 & "\.lenv"
 
+if exist %~dp0 & "\.lenv\" (
+    .lenv\Scripts\activate.bat
+) else (
+    echo "Setting up Lunar Virtual Environment for the first time."
+    python3 -m venv .lenv
+    .lenv\Scripts\activate.bat
+    pip install -r requirements.txt
+)
 
+echo "Starting Lunar."
 set FLASK_APP=lunar.py
+flask run
